telegram/requirements.txt
python-dotenv>=1.0.1,<1.1.0
requests>=2.32.0,<3.0.0
